backend/core/tests.py
"""
Core Tests Module

Comprehensive test suite for:
- Model validation
- API endpoints
- Business logic
- Authentication
- Permissions
- Edge cases

Features:
- Unit tests
- Integration tests
- API tests
- Mock objects
- Test fixtures
- Coverage reporting
"""

from django.test import TestCase
from rest_framework.test import APITestCase
from django.contrib.auth.models import User
from .models import Category, Product, Stock

class CategoryTests(TestCase):
    """
    Category Model Tests
    
    Tests category functionality:
    - Creation
    - Validation
    - Relationships
    - Edge cases
    """
    def setUp(self):
        """Set up test data"""
        self.category = Category.objects.create(name="Test Category")

    def test_category_creation(self):
        """Test category creation and validation"""
        self.assertEqual(self.category.name, "Test Category")

class APITests(APITestCase):
    """
    API Endpoint Tests
    
    Tests API functionality:
    - Authentication
    - CRUD operations
    - Permissions
    - Response formats
    """
    # Test implementations... 