backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
